// generated/src/app/api-models/pmb-subsc-snapshot-arc.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbSubscSnapshotArc {
  seqSubsId: number;
  seqSubsUpdate: number;
  changeDateTime: Date;
  effectiveDate: Date;
  seqGroupId: number;
  planCode: string;
  familySize: number;
  spouseFlag: string;
  otherStatusFlag: string;
  premOverrideStep: number;
  premOverrideAmt: number;
  premOverrideCode: string;
  termDate: Date;
  subscriberDob: Date;
  subscriberGender: string;
  salary: number;
  medicareStatus: string;
  subscDept: string;
  subscLocation: string;
  billEffectiveFrom: Date;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  rateType: string;
}